changed level file so -> trunk runnable

+++ code/trunk/src/orxonox/objects/SpaceShip.h
@@ -83,7 +83,7 @@ namespace orxonox
             float mouseX_;
             float mouseY_;
 
-            int emitterRate_;
+            float emitterRate_;
     };
 }
